es 1.0.1 doesnt implicitly convert lat and lng strings to int

+++ backend/feed.py
@@ -105,8 +105,8 @@ class Feed(object):
         sortby.append({
                                 "_geo_distance" : {
                                               "coord" : {
-                                                             "lat" : int(coord[0]),
-                                                            "lon" : int(coord[1])
+                                                             "lat" : float(coord[0]),
+                                                            "lon" : float(coord[1])
                                                     },
                                             "order" : "asc",
                                             "unit" : "km"
@@ -129,8 +129,8 @@ class Feed(object):
                 "geo_distance" : {
                     "distance" : "{0}km".format(radius),
                     "coord" : {
-                        "lat" : int(coord[0]),
-                        "lon" : int(coord[1])
+                        "lat" : float(coord[0]),
+                        "lon" : float(coord[1])
                     }
                 }
             } 
@@ -162,8 +162,8 @@ class Feed(object):
         sortby.append({
                                 "_geo_distance" : {
                                               "coord" : {
-                                                             "lat" : int(coord[0]),
-                                                            "lon" : int(coord[1])
+                                                             "lat" : float(coord[0]),
+                                                            "lon" : float(coord[1])
                                                     },
                                             "order" : "asc",
                                             "unit" : "km"
@@ -191,8 +191,8 @@ class Feed(object):
                 "geo_distance" : {
                     "distance" : "{0}km".format(radius),
                     "coord" : {
-                        "lat" : int(coord[0]),
-                        "lon" : int(coord[1])
+                        "lat" : float(coord[0]),
+                        "lon" : float(coord[1])
                     }
                 }
             }
